Implemented automatic linking when building with Visual C++. Unlike other projects' use of automatic linking, this feature is optional. It is enabled by defining CPPDOM_AUTO_LINK and/or CPPDOM_DYN_LINK. If CPPDOM_AUTO_LINK is defined without CPPDOM_DYN_LINK, then the static version of the debug or release library is used for linking. Otherwise, the dynamic version is used.
The reason that I made this optional is largely for backwards compatibility
with existing uses of CppDOM. If automatic linking were not optional, a
user-level build that was not updated to define CPPDOM_DYN_LINK would end
up linking against both the static and dynamic library, thus resulting in
multiply defined symbols at link time. Removing the test to determine if
CPPDOM_AUTO_LINK is defined would result in this library behaving the way
that all others that I have seen that use automatic linking.


git-svn-id: https://xml-cppdom.svn.sourceforge.net/svnroot/xml-cppdom/branches/releng-0-6@517 bdc45a6f-1a19-0410-8385-c3db3a2b84f1